[Icons] Algorithmically generate icons if not provided
Underlying issue for icon revampment is inability to
easily distinguish objects of different types if they
do not have custom icons. Adding custom icons is the
best part of the solution, but this additionally adds
procedurally generated icons based on component type
hash codes which emulate the general look of the
default component. These do not clearly indicate
what type of thing a thing is, but do make it easy to
identify when two things are or are not of the same
type.