[Inspector] Make view-setting part of ViewProvider
Add the ability to set contained views to the
ViewProvider interface (an appropriate
corollary to its existing get method). This
method may reject changes - by default, View
simply returns false to indicate that the
change has been ignored.

This is to facilitate the decoupling of the
Inspector view from the control used to
switch views (view switching is a
cross-cutting concern, occurring in multiple
Inspector implementations and desirable in a
fiew contexts.)

This ultimately serves to support combo box
style switcher described in nasa/mct#135