[Inspector] Add test for default view provider
Making the view switcher a custom view type
allows reuse in multiple plug-ins (in their
custom inspectors), but if this view type is
missing then those plug-ins will fail. This
unit test ensures that this type of view is
provided by the platform.

Since several other view types are similarly
fundamental, their availability is also
tested in the same manner.