[Ownership] Prevent ownership change on bootstrap comps
In nasa/mct#123 the user is able to change ownership
of My Sandbox, which means they subsequently cannot
edit their own sandbox. Worse, when they restart MCT
they see no sandbox at all.

This adds a check to prevent this. When ownership
mutability is being determined, consult the list
of bootstrap components - this is the list of
components seen by the user in the User Environment.
Categorically disallow changes of ownership to any
of these (this includes both My Sandbox, and any
future top-level components introduced, as these
are presumed to be inherently tied to the user. An
example of a top-level component not yet introduced
is User Tags, described in nasa/MCT-Plugins#39.)