[Tests] Fix scatter plot test
Scatter plot test was failing on auto-expand
behavior after merge with new controls for
non-time non-time plots. This was due to the
associated bounds managers, which feature an
optimization based on the assumption that
data will not be received (by the plot) in
the middle of an existing set (only prepend
or append is expected.)

As this is currently an accepted limitation
of plots, unit tests are amended to ensure
that data which is expected to expand the
plot includes time stamps which have not
been previously exercised.