[Views] Disregard view type when testing info equality
Certain set operations upon ViewInfo objects
are contingent upon the treatment of ViewInfo
objects describing the same class as equal,
even if they apply to different view types.
Failure to exhibit this sort of equality is
the root cause of nasa/mct#195.

This change removes consideration of view
type and hashCode from equals (this was
added erroneously in a previous commit
triggering the regression.)