[Tests] Modify spinner test for Xvfb compatibility
TestMenuItemSpinner was failing under Xvfb
(but not other environments). Appears that
cause is some difference (possibly in layout)
regarding the way spinners respond to clicks
from FEST under that environment.

Test is modified such that clicks are now
directed explicitly to the text field in
the spinner, such that subsequent typing
exercises the intended behavior.