[Launcher] Pop up user dialog lazily
When MCT is launched with no user specified,
the DefaultIdentityManager should pop up a
dialog of available users. However, the
persistence service is used to determine
which users are available, and this service
not necessarily been created at the time the
default identity manager is created (due
to bundle ordering.) When persistence is
null, nasa/mct#199 occurs.

In this change, default identity manager
does not try to identify the current user
until the first time getCurrentUser is
invoked. This occurs from LaunchMCTService,
which depends upon the PersistenceService
having been bound (so it is guaranteed to
be available.)