[Info View] Add class for custom visual controls
Adds abstract class CustomVisualControl,
which plugins can implement fully if
they wish to expose custom controls in
the Info View. This supports nasa/mct#192
which in turn block nasa/MCT-Plugins#32.

The inclusion of this class in mctcore
is somewhat contradictory to the ongoing
maintenance issue nasa/mct#155, but is
consistent with current packaging. As
future work it may be desirable to move
the full gov.nasa.arc.mct.gui package
out of mctcore and into a Swing-specific
bundle; the addition of this class is
not expected to add any difficult to
that refactoring.