[Persistence] Nullify lastModified before update
Set lastModified to null before persisting,
but specify that it is not nullable (such
that CURRENT_TIMESTAMP is used instead.)
This ensures that lastModified always uses
the database's version of the current time
and always records this at the time the
object is updated. Object polling also
uses CURRENT_TIMESTAMP from DB, so if this
is not updated or recorded from different
clocks then component updates may be missed,
resulting in nasa/mct#110 and nasa/mct#134.