[Persistence] Evict cached children on update from database
Referenced component (or, more specifically,
ComponentSpec object) collections are stored
in the L2 cache as of #48; this caching is
useful in reducing latency related to repeated
loading of referenced components, particularly
where the database is accessed over a network.
However, the L2 cache is not refreshed by
changes to the parent ComponentSpec, causing
child references to fall out of date, which
results in nasa/mct#134 and some cases of
nasa/mct#110.

This fixes the issue by evicting these
collections explicitly for any components
which are updated from the database. This
allows unchanged components to continue to
use the cache, retaining the performance
benefits when viewing objects.

The explicit references to Hibernate involved
introduce a dependency which may not be
necessary; this is addressed by followup
issue nasa/mct#212.