[Delete] Enable delete all behavior
Invoke the descendant probing algorithms to
assemble lists of components which can be
deleted, which cannot be deleted but would
be removed, and cannot be removed, and
complete the action if it can be completed
without violating the latter.

Dialogs pop up conditionally, but do not
present complete information desired.