[Delete] Use Window Manager for Delete All dialog
Use the Window Manager (as opposed to static methods
of OptionBox) to show the warning dialogs related
to the Delete All action introduced for nasa/mct#89.

This slightly reduces the explicit Swing
dependencies associated with the action and
facilitates automated testing of interactions
(the WindowManager can be mocked, but the static
methods of OptionBox cannot). These are ongoing
concerns of nasa/mct#155.