[Directory] Defer tree expansion details to node
Tree expansion triggers loading of components on
a background thread (with placeholder) as this
may potentially take some time when there are a
lot of children. Prior behavior was somewhat
redundant to (and not complementary to) the
exisiting code to handle node update in
NodeViewManifestation's updateMonitoredGUI.

Initial testing indicates that ordering issues
between these calls results in the node collapse
observed in nasa/mct#106. Invoking
updateMonitoredGUI to handle this behavior
appears to resolve the issue.