[Assembly] Use mysql connector from lib folder
In 1d79da61212d92462070001c9b0b83ffb2574135
database creation script was modified to
include a classpath reaching the mysql
connector pointing to the ext folder, in
order to reflect the move there.

This connector was moved again to the
database persistence bundle in order to
support nasa/mct#109 (treating ext folder
as Java extensions was used to circumvent
OSGi classpath isolation, but this required
extra command line parameters.)

MySQL connector reintroduced to lib folder
to support resolution of nasa/mct#221.
This change points the database creation
script back to lib (as opposed to ext)
in order to locate this connector.