[Database] Add dependency to Platform service
Add a dependency (via declarative services) to
PersistenceProvider, upon Platform. This prevents
the PersistenceProvider from being instantiated
before the platform is available; this also
means that platform's activator may be started
before the database is started. This allows for
a database startup timer to start before getting
stuck in a long DB wait, which allows for user
feedback regarding nasa/mct#225.