[Persistence] Add Step-Behind Cache
Introduces a 'Step-Behind Cache' to handle
custom queries. This returns values which are
effectively one step behind the results given
by the real operation being performed.

This is used to provide quick answers for
custom database lookups (such as getting
all users) while still receiving updates
eventually, without incurring the cost of
polling. This is useful in cases (such as
getting a list of users) where the data
changes infrequently, and where it is
acceptable to be somewhat out-of-step with
the ground truth.

This addresses the remaining known cause
of latency in nasa/mct#111.