[Persistence] Use thread pool for step-behind cache
Use a thread pool in step-behind cache to avoid creating an
excessive number of background threads.

While the existing time delay provides some throttling
with regard to the number of actual look-ups initiated, it
is not a guarantee against rampant thread creation,
particularly where time delay is less than the time taken
to complete the underlying lookup operation.