[Persistence] Keep bootstrap cache up-to-date during start up
Fix for nasa/mct#111 involves a "step-behind cache"
for bootstrap components which is sometimes behind
the latest result from the database. This is
acceptable during operation or reuse of MCT, when
the set of bootstrap components changes very
rarely and is a background delay is unlikely to
be noticed. However, during startup of MCT on a
new database (particularly when configured to
automatically add users, or when plugins introduce
new bootstrap components) this can lead to errors
which emerge from the failure to recognize
bootstrap components just after they were
introduced.

Addressed here:

* Pre-populate the bootstrap cache once when
  the persistence service is activated.
* Explicitly the cache whenever components are
  tagged as bootstraps from the running
  instance of MCT.