[Save All] Allow Save without ObjectManager
Consolidating Save and Save All behavior
into one action, for nasa/mct#131.

Save All (which has become the consolidated
Save) previously required an ObjectManager
capability to be present, as the presence
thereof was what distinguish Save situations
from Save All situations. This change allows
for the ObjectManager capability to be
unavailable, in which case the action
functions as te per-object save.

Additionally, rules for action eligibility
are relaxed to allow Save when not all
objects are writeable (e.g. because someone
else owns the object.) Objects which cannot
be written are not saved.