[Platform] Avoid NPE on getMySandbox
Remove risk of hitting NPE on getMySandbox in
PlatformImpl. Stored reference to sandbox is
obtained during call to getBootstrapComponents,
but if that has not been called when getMySandbox
is needed, an NPE occurs.

This is significant for nasa/MCT-Plugins#134,
which requires some deployment-specific additions
to My Sandbox; since the most obvious place to
do this is at bootstrapping time, e.g. before
or during the first call to PlatformImpl's
getBootstrapComponents.