[Persistence] Refresh bootstrap cache when creating user
When introducing a new user to the database
(which includes adding My Sandbox as a bootstrap
component) refresh the list of bootstrap
component identifiers. Otherwise, the next
call(s) to retrieve bootstraps may not reflect
this change; this results in the apparent
absence of My Sandbox on first run see in
nasa/mct#245.