[Canvas] Avoid null TransferHandler
Transfer handlers on Inspector labels
are not necessary after inspected view
is clear, as there is nothing to transfer.
However, Swing's event-handling does go
back to inform the original transfer
handler that an event has completed; if
this happens after the component's
transfer handler has been nullified,
nasa/mct#172 results.

To avoid this issue, instead of nullifying
the transfer handler, simply introduce
checks such that it becomes inert when
there is no view in the Inspector.